requirements.txt
numpy
matplotlib
pettingzoo
torch==1.13.1
scikit-learn
scipy
rlcard
pygame
pyro-ppl
typing